pyproject.toml
[project]
name = "cloud-compute-mcp-server"
version = "0.1.0"
description = "Cloud Compute MCP server"
readme = "README.md"
requires-python = "~=3.11"
dependencies = [
    "mcp[cli]==1.9.4",
    "pyyaml==6.0.2",
    "requests==2.32.4",
]

[dependency-groups]
dev = [
    "ipykernel==6.29.5",
    "jupyter==1.1.1",
    "nbformat==5.10.4",
]
